chore: add license
This sets the license for this proeject as the Functional Source
License. The reason for setting that over the MIT License is because I
am not sure whether I plan to try and monetize from this website or not
but I want to keep that option open. If I decide to not monetize and
keep this completely open, then I may retroactively change the license
though I am not sure of the legal aspects of that.