chore: run pre-commit checks manually if in nix shell
This runs the flake checks if we are not in a nix shell, but it
manually runs the pre-commit checks if we are in a nix shell. The
reason for this decision is:

1. `nix build` is too slow to my liking and it becomes annoying
especially when a series of small commits are made.

2. Also, it does not allow filtering out the changed files and running
`treefmt` only on those files.

This does add a maintenance overhead in that both the nix checks being
ran and the script being ran here need to kept in sync. That should be
fine since these scripts are unlikely to be changing that frequently.
Also, mismatches should be caught in CI where all the nix checks should
be ran.