ci(gh-actions): approve the PR based on a comment (#2)
Github does not allow the author of a PR to be a reviewer which means
it is not possible to enforce at least a single review. So, as a
workaround, allow the actions bot to mark the PR as approved based on
a comment.

REFERENCE: https://github.com/orgs/community/discussions/6292