feat: seed database (#15)
* refactor: alias database type

* fix: make the `CHECK` constraint deterministic

`sqlite` does not allow non-determinstic expressions within `CHECK`
constraints so this just hardcodes the maximum value. This should be
fine since:
    - the maximum year value will only change once per year
    - it is okay if the year is not updated immediately when the year
      changes

Also, a test can/will be added to ensure that a question paper with the
latest year exists so that the constraint can be updated when a new year
comes (the test should fail at that point).

* feat: seed database

* test: create test database

* test: test seeding

* fix: use 'union all'

* test: check current year present in question_papers

* test: hardcode path to schema file